# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all members and guests
      parameters:
        - name: status
          in: query
          required: false
          description: Filter by record status
          schema:
            type: string
            enum: [member, guest, all]
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createGuest
      tags: [members]
      summary: Register a guest (informational record, no login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestInput"
      responses:
        "201":
          description: Guest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /members/import:
    post:
      operationId: importMembers
      tags: [members]
      summary: Bulk import members from CSV rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a member by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member's attributes (role, categories, formation track, contact)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Remove a member or guest record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
  /members/{id}/promote:
    post:
      operationId: promoteGuest
      tags: [members]
      summary: Promote a guest to member (reuses existing profile)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoteInput"
      responses:
        "200":
          description: Promoted member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /discipleship:
    get:
      operationId: listDiscipleships
      tags: [discipleship]
      summary: List all discipleship relationships
      responses:
        "200":
          description: List of relationships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discipleship"
    post:
      operationId: createDiscipleship
      tags: [discipleship]
      summary: Create a discipler to disciple relationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscipleshipInput"
      responses:
        "201":
          description: Relationship created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discipleship"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /discipleship/{id}:
    patch:
      operationId: updateDiscipleship
      tags: [discipleship]
      summary: Update discipleship progress notes or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscipleshipUpdate"
      responses:
        "200":
          description: Updated relationship
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discipleship"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteDiscipleship
      tags: [discipleship]
      summary: Remove a discipleship relationship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    OkResult:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
    Member:
      type: object
      description: Member row. Contact/formation fields are only populated for leader/auxiliary viewers.
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [member, guest]
        role:
          type: ["string", "null"]
          enum: [leader, auxiliary, member, null]
        categories:
          type: array
          items:
            type: string
            enum: [host, discipler, disciple]
        formationTrack:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
          description: Birth date (YYYY-MM-DD)
        avatarPath:
          type: ["string", "null"]
          description: Object path of the profile picture
        invitedBy:
          type: ["string", "null"]
        joinedAt:
          type: ["string", "null"]
        active:
          type: boolean
      required:
        - id
        - name
        - status
        - categories
        - birthDate
        - avatarPath
        - active
    MemberDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [member, guest]
        role:
          type: ["string", "null"]
          enum: [leader, auxiliary, member, null]
        categories:
          type: array
          items:
            type: string
            enum: [host, discipler, disciple]
        formationTrack:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
          description: Birth date (YYYY-MM-DD)
        avatarPath:
          type: ["string", "null"]
          description: Object path of the profile picture
        invitedBy:
          type: ["string", "null"]
        joinedAt:
          type: ["string", "null"]
        active:
          type: boolean
        disciples:
          type: array
          items:
            $ref: "#/components/schemas/Discipleship"
        disciplers:
          type: array
          items:
            $ref: "#/components/schemas/Discipleship"
      required:
        - id
        - name
        - status
        - categories
        - birthDate
        - avatarPath
        - active
        - disciples
        - disciplers
    GuestInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
        invitedBy:
          type: string
      required:
        - name
    PromoteInput:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email
    MemberUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [leader, auxiliary, member]
        categories:
          type: array
          items:
            type: string
            enum: [host, discipler, disciple]
        formationTrack:
          type: string
        birthDate:
          type: ["string", "null"]
          description: Birth date (YYYY-MM-DD), null clears it
        avatarPath:
          type: ["string", "null"]
          description: Object path of the profile picture, null clears it
        active:
          type: boolean
    MemberImportInput:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/MemberImportRow"
      required:
        - rows
    MemberImportRow:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        formationTrack:
          type: string
      required:
        - name
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
      required:
        - imported
        - skipped
...
    Discipleship:
      type: object
      description: |
        Each side is either an internal member (id set) or an external person
        from another Life Group (external name set). At least one side is
        internal. disciplerName/discipleName always carry the display name.
      properties:
        id:
          type: integer
        disciplerId:
          type: ["integer", "null"]
        disciplerName:
          type: string
        externalDisciplerName:
          type: ["string", "null"]
          description: Free-text name when the discipler is from another Life Group
        discipleId:
          type: ["integer", "null"]
        discipleName:
          type: string
        externalDiscipleName:
          type: ["string", "null"]
          description: Free-text name when the disciple is from another Life Group
        startDate:
          type: string
        status:
          type: string
          enum: [active, paused, completed]
        notes:
          type: ["string", "null"]
      required:
        - id
        - disciplerId
        - disciplerName
        - externalDisciplerName
        - discipleId
        - discipleName
        - externalDiscipleName
        - startDate
        - status
    DiscipleshipInput:
      type: object
      description: |
        For each side provide exactly one of the internal id or the external
        name. At least one side must be an internal member.
      properties:
        disciplerId:
          type: integer
        externalDisciplerName:
          type: string
          minLength: 1
        discipleId:
          type: integer
        externalDiscipleName:
          type: string
          minLength: 1
        notes:
          type: string
    DiscipleshipUpdate:
      type: object
      description: |
        Sides may be reassigned; when changing a side, provide exactly one of
        the internal id or the external name for it. At least one side must
        remain an internal member.
      properties:
        disciplerId:
          type: ["integer", "null"]
        externalDisciplerName:
          type: ["string", "null"]
        discipleId:
          type: ["integer", "null"]
        externalDiscipleName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, paused, completed]
        notes:
          type: string
...