# lib/api-zod/src/generated/types/memberDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Discipleship } from './discipleship';
import type { MemberDetailCategoriesItem } from './memberDetailCategoriesItem';
import type { MemberDetailRole } from './memberDetailRole';
import type { MemberDetailStatus } from './memberDetailStatus';

export interface MemberDetail {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  status: MemberDetailStatus;
  /** @nullable */
  role?: MemberDetailRole;
  categories: MemberDetailCategoriesItem[];
  /** @nullable */
  formationTrack?: string | null;
  /**
     * Birth date (YYYY-MM-DD)
     * @nullable
     */
  birthDate: string | null;
  /**
     * Object path of the profile picture
     * @nullable
     */
  avatarPath: string | null;
  /** @nullable */
  invitedBy?: string | null;
  /** @nullable */
  joinedAt?: string | null;
  active: boolean;
  disciples: Discipleship[];
  disciplers: Discipleship[];
}