# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemberUpdateCategoriesItem } from './memberUpdateCategoriesItem';
import type { MemberUpdateRole } from './memberUpdateRole';

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  phone?: string;
  role?: MemberUpdateRole;
  categories?: MemberUpdateCategoriesItem[];
  formationTrack?: string;
  /**
     * Birth date (YYYY-MM-DD), null clears it
     * @nullable
     */
  birthDate?: string | null;
  /**
     * Object path of the profile picture, null clears it
     * @nullable
     */
  avatarPath?: string | null;
  active?: boolean;
}